6.0
6.0 A'zadi Update is out! @everyone
We are thrilled to unveil the much-anticipated first major update to our game, marking an extensive journey of development. Introducing the A'zadi region, a captivating locale renowned for its unwavering reverence for the Kon'chi deities and a deep-rooted reluctance towards external interactions. Players will immerse themselves in the intricate dynamics of the empire, navigating the complex societal structures as factions debate the pivotal choice between maintaining isolation or embracing global engagement. Explore the rich tapestry of A'zadi's culture, confront challenging dilemmas, and shape the destiny of a civilization on the brink of profound change. Welcome to a world where tradition clashes with progress, where every decision carries weight, and where the future hangs in the balance.

This update includes:
A new region
15 new starting nations
2 new form-able nations
The complex voting system of the Amma Narsa
The brand new religion of Kon'chi
New Provinces
Greater localisations
Many bug fixes